# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API publique — livraison
  version: 1.0.0
  description: |
    API des e-commerçants : créer et suivre des courses depuis une boutique en ligne,
    un logiciel de caisse ou une application indépendante.

    ## Authentification
    Créez une clé dans l'application marchand (**Profil › Intégrations**). Elle n'est affichée qu'une fois.
    Envoyez-la à chaque requête : `Authorization: Bearer lv_xxxxxxxx_…` (ou en-tête `X-Api-Key`).

    Portées : `orders:read` (lire les courses et le point), `orders:write` (créer, annuler),
    `stock:read` (produits et stock). Les zones, entrepôts et devis sont accessibles à toute clé.

    ## Idempotence
    `POST /orders` exige l'en-tête `Idempotency-Key` (par exemple un UUID par commande). Une requête
    relancée avec la même clé renvoie la réponse d'origine (en-tête `Idempotent-Replayed: true`)
    au lieu de créer un doublon. Les clés sont conservées 24 heures.

    ## Limitation de débit
    120 requêtes par minute et par clé (en-têtes `X-RateLimit-Limit` et `X-RateLimit-Remaining`).
    Au-delà : réponse `429`, réessayez après le délai indiqué par `Retry-After`.

    ## Erreurs
    Toujours au format `{ "message": "…", "errors": { "champ": ["…"] } }`. `401` clé absente ou invalide,
    `403` portée manquante ou compte suspendu, `404` course introuvable, `422` données invalides ou règle métier.

    ## Webhooks
    Déclarez une adresse HTTPS dans **Profil › Intégrations** pour recevoir les événements
...
servers:
...
security:
  - bearer: []
...
paths:
...
  /orders/{tracking_code}:
    get:
      tags: [Courses]
      summary: Suivre une course (avec son historique)
      parameters:
        - { name: tracking_code, in: path, required: true, schema: { type: string, example: LV-7K2M-QX9P } }
      responses:
        '200':
          description: Course et événements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Order' }
        '404': { description: 'Course introuvable' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Clé API (lv_…)
...
  schemas:
...
    Status:
      type: string
      enum: [pending, confirmed, rejected, pickup_assigned, pickup_in_progress, picked_up, at_hub, delivery_assigned, out_for_delivery, delivered, delivery_failed, rescheduled, return_assigned, returning, returned, cancelled, lost]
...
    Order:
      type: object
      properties:
        tracking_code: { type: string, example: LV-7K2M-QX9P }
        merchant_reference: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/Status' }
        status_label: { type: string, example: En chemin }
        tracking_url: { type: string, format: uri, description: 'Page de suivi à envoyer au client' }
        delivery_code: { type: string, description: 'Code à donner au client, demandé par le livreur à la livraison' }
        recipient:
          type: object
          properties:
            name: { type: [string, 'null'] }
            phone: { type: string }
            phone2: { type: [string, 'null'] }
        delivery:
          type: object
          properties:
            zone_id: { type: integer }
            zone_name: { type: string }
            address: { type: [string, 'null'] }
            landmark: { type: [string, 'null'] }
            scheduled_date: { type: [string, 'null'], format: date }
        from_warehouse: { type: boolean }
        description: { type: [string, 'null'] }
        items:
          type: array
          items:
            type: object
            properties:
              product_id: { type: [integer, 'null'] }
              label: { type: string }
              quantity: { type: integer }
              unit_price: { type: integer }
        amounts:
          type: object
          properties:
            items_amount: { type: integer }
            delivery_fee: { type: integer }
            fee_payer: { type: string, enum: [merchant, recipient] }
            cod_amount: { type: integer, description: 'Total à encaisser auprès du client' }
            collected_amount: { type: [integer, 'null'] }
            shipping_fee: { type: [integer, 'null'] }
        attempts_count: { type: integer }
        last_incident: { type: [string, 'null'] }
        events:
          type: array
          description: Présents sur GET /orders/{tracking_code}
          items:
            type: object
            properties:
              type: { type: string }
              status: { type: [string, 'null'] }
              note: { type: [string, 'null'] }
              incident: { type: [string, 'null'] }
              rescheduled_to: { type: [string, 'null'], format: date }
              at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        delivered_at: { type: [string, 'null'], format: date-time }
        cancelled_at: { type: [string, 'null'], format: date-time }
        returned_at: { type: [string, 'null'], format: date-time }
...